requirements.txt

africastalking==1.2.3
asgiref==3.4.1
boto3==1.21.0
Django==3.1.4
django-cors-headers==3.4.0
django-filter==2.3.0
django-imagekit==4.0
django-multiselectfield==0.1.12
django-oauth-toolkit==1.2.0
django-background-tasks==1.2.5
django-rest-framework-social-oauth2==1.1.0
django-storages==1.12
djangorestframework==3.12.2
djangorestframework-bulk==0.2.1
djangorestframework-gis==0.15
git+https://github.com/michameiu/drf-autodocs.git@686a788ec0d98986566688a2badf4bc2a698e1ab#egg=drf_autodocs

django-sistch-drf-stats@git+https://github.com/sisitech/django-common-stats.git@23d9098669a542a5c3fd15ba76b06b60f8cfea4f
django-user-agents==0.4.0
gunicorn==20.0.4
oauthlib==3.1.0
onesignal-sdk==2.0.0
openpyxl==3.0.4
pilkit==2.0
Pillow==9.1.0
psycopg2-binary==2.9.5
pydot==1.4.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
sendgrid==6.4.7
six==1.15.0
slack-webhook==1.0.4
social-auth-app-django==4.0.0
social-auth-core==3.3.3
urllib3==1.25.9
user-agents==2.1
zipp==3.1.0
weasyprint==59.0
lxml==4.8.0
python-memcached==1.59
python-dotenv==1.0.0
